+++ wcfsetup/install/files/lib/acp/form/LanguageImportForm.class.php
@@ -38,22 +38,10 @@ class LanguageImportForm extends AbstractForm {
 	public $filename = '';
 	
 	/**
-	 * import field
-	 * @var	string
-	 */
-	public $importField = 'languageFile';
-	
-	/**
 	 * language object
-	 * @var	\wcf\data\language\Language
+	 * @var	Language
 	 */
-	public $language = null;
-	
-	/**
-	 * import language file
-	 * @var	string
-	 */
-	public $languageFile = '';
+	public $language;
 	
 	/**
 	 * list of available languages
@@ -62,20 +50,27 @@ class LanguageImportForm extends AbstractForm {
 	public $languages = [];
 	
 	/**
+	 * source language object
+	 * @var	Language
+	 */
+	public $sourceLanguage;
+	
+	/**
+	 * source language id
+	 * @var	integer
+	 */
+	public $sourceLanguageID = 0;
+	
+	/**
 	 * @inheritDoc
 	 */
 	public function readFormParameters() {
 		parent::readFormParameters();
 		
-		// import
-		if (isset($_POST['languageFile']) && !empty($_POST['languageFile'])) {
-			$this->languageFile = $_POST['languageFile'];
-			$this->filename = $_POST['languageFile'];
-		}
 		if (isset($_FILES['languageUpload']) && !empty($_FILES['languageUpload']['tmp_name'])) {
-			$this->importField = 'languageUpload';
 			$this->filename = $_FILES['languageUpload']['tmp_name'];
 		}
+		if (isset($_POST['sourceLanguageID'])) $this->sourceLanguageID = intval($_POST['sourceLanguageID']);
 	}
 	
 	/**
@@ -86,7 +81,17 @@ class LanguageImportForm extends AbstractForm {
 		
 		// check file
 		if (!file_exists($this->filename)) {
-			throw new UserInputException('languageFile');
+			throw new UserInputException('languageUpload');
+		}
+		
+		if (empty($this->sourceLanguageID)) {
+			throw new UserInputException('sourceLanguageID');
+		}
+		
+		// get language
+		$this->sourceLanguage = LanguageFactory::getInstance()->getLanguage($this->sourceLanguageID);
+		if (!$this->sourceLanguage->languageID) {
+			throw new UserInputException('sourceLanguageID');
 		}
 		
 		// try to import
@@ -96,10 +101,10 @@ class LanguageImportForm extends AbstractForm {
 			$xml->load($this->filename);
 			
 			// import xml document
-			$this->language = LanguageEditor::importFromXML($xml, -1);
+			$this->language = LanguageEditor::importFromXML($xml, -1, $this->sourceLanguage);
 		}
 		catch (SystemException $e) {
-			throw new UserInputException($this->importField, $e->getMessage());
+			throw new UserInputException('languageUpload', $e->getMessage());
 		}
 	}
 	
@@ -120,12 +125,21 @@ class LanguageImportForm extends AbstractForm {
 	/**
 	 * @inheritDoc
 	 */
+	public function readData() {
+		parent::readData();
+		
+		$this->languages = LanguageFactory::getInstance()->getLanguages();
+	}
+	
+	/**
+	 * @inheritDoc
+	 */
 	public function assignVariables() {
 		parent::assignVariables();
 		
 		WCF::getTPL()->assign([
 			'languages' => $this->languages,
-			'languageFile' => $this->languageFile
+			'sourceLanguageID' => $this->sourceLanguageID
 		]);
 	}